RO-29
Create basic script for camera behavior. finally understand the values
im munipulating. will settle for a simple zoom in and zoom out. code
currenty zooms in and out as player one enters the fore or background,
this will be adjusted further today to better position how the camera
zooms in now that the triggers for when to zoom in or out are now in
place.