Input delay
attacks now come out on an adjustable input delay. opponenet now goes
flying when hit. Need to better fine tune this as player does not always
travel along expected path. Camera locks on to hit player but this
happens so fast we may want to make a more smooth transition as the
player is flying through the air