RO-170
removed switch statement from player. adjusted some of the code logic to
assign players current state via the animations played. This can be
built upon further using the animator to assign scripts to animation
state if we continue with this implementation. for now however anytime
an animation call is needed for a specific component a reference to the
players PlayerAnimator will be required.

Note: I also duplicated the PlayerOne in the scene and disabled the
original PlayerTwo.