RO-134
addressed momentum bar still allowing player to as notated in Test Case
2-84 -2
added a "isTimer" bool to the MomentumBar.Cs class. the reset bar method
now checks if the bar has hit its reset point in which it will set
isTimer to false. The OnHit function now only executes when isTimer is
set to false. We now set isTimer to true at the moment we activate
isHype in the isMaxed() method