RO-140
added additional BlockHit animation to PlayerANim.cs
Added a disable hitbox method in collidewithopponent.cs component. This
method sets a isblocked bool to true I came up with this solution due to
the player colliding with the blocking area hitbox and then the
opponents reguar hitbox, so now if the player collides with a block
hitbox first, even if collides with the body, our switch will have been
triggered. we then set the bool to false once it has touched the body.
im not totally sure how the hell my solution works but i know it does,
my brain is a little fried right now lol. EIther way I need to test to
see if the opponent can be hit from behind even while blocking next. and
then i still need to resize the hitboxes for both players.