RO-154
Added an additional CloseAttack-esque method to the player anim class.
This method may be a code smell as it is repeating behavior shared with
the regular close attack and attack animations. this can possibly be
refactored with the use of a bool tracking a hype attack and normal
attack on the player.cs but that can be discussed or fixed later, for
now despite the code smell the expected behavior is in place. Ready for
testing against this branch