RO-154
moved the set hype false coroutine to the Hitbox.cs. Since the animation
is what calls the hitbox script to open and fire off, i have the setting
of the hype attack logic happening the moment the attack hitbox is open.
which should mean it only resets when the animation plays as this is the
only code called during its animation