RO-154
polishing up some small refactors before merging into AlfredNewBranch.
added a RingOut() method to the Movement.cs class, now when the ringout
effect happens, the movement script handles the velocity changes to
transition the player out the ring. currently looks a lot smoother than
before however the camera has trouble keeping up with the opponenet on
the initial hit. this could be rectified in the future by adding a hit
pause at the moment of impact before the player is set flying