RO*336
played around with the Animator Controller in Unity, added a few
additional transitions between various animation clips to seg way into
jump. added a .30sec delay before setting the IsAttacking bool in the
Playeranim.cs to true. will probably need to refactor how we allow the
next input command for the attack sequence as it is now autofiring the
second attack because the isAttacking bool is left active too long but a
better solution would be to refactor somewhere else possibly where we
can dictate the combo sequence flow like in the combat.cs class most
likely or the Hitbox.cs class or maybe even adjusting to check for
button down and tying that to when we will allow next input