336
made a lot of additional refactors and changes. I cant even remember
everything lol but the entire PlayerAnim.cs has been replaced with a
more SOC friendly AnimationManager. I tried to get rid of the web of
animatior calls but I ended up mostly going back to that way of
connecting animations as it was not producing the immediate expected
results trying to call the animations directly. I ended up using a lot
of CoRoutines but its my understanding that can get pretty taxing. we
will have to see how it effects things as we get deeper into
development. For now the combat feels a little bit better however I
still not satisfied with the fluidness of the jump and its animation
transitions itself. All that being said the jump attack does seem to be
executing decently, all that is left is to connect the hitboxes to it
and call the hitbox functions on the jump attack animation.