RO-319
Added additional methods to combat.cs and created a new Combat Trigger
just for the Hype Attack Hitbox. Additionall I repurposed the
JumpAttackHitbox and made it the hitbox that the hypeattack logic
triggers for hit detection. Added a hype attack delay before resetting
the hype attack bool, this plays for the duration of the hype attack but
can be adjusted. Tweaked jump attack hitbox to cover the area of marie o
netts mallet during its animation. I also increased the size of the
mallets hitbox on all animations so it now covers a more gameplay
experience approiate point of contact. renamed JumpAttackHitbox to
HypeAttackHitBox